Propagating IOException thrown by char stream
Currently CharStream throws IOException to indicate end of stream
and that exception is swallowed by the generated token manager.
But this is terribly awful design decision. There may be many
valid reasons for java.io.Reader or java.io.InputStream to throw
IOException.
I am working on making thrown IOException visible to the users
of the generated token manager and parser.
The following commits will contain refactorings that update CharStream
API to return meaningful value on end of stream.